+++ src/commands/mc.workspace.importStructure.ts
@@ -1,24 +1,39 @@
 import { QuickPickItem, window } from "vscode";
 import { CurrentPackProvider, PackItem, PackItemType } from "../class/CurrentPackProvider";
 import { join } from "path";
-import { copyFile, copyFileSync, readdirSync } from "fs";
+import { copyFile, copyFileSync, existsSync, mkdir, mkdirSync, readdirSync } from "fs";
 import { notifications } from "../class/NotificationProvider";
 import assert = require("assert");
+import { UtilFunctions } from "../class/UtilFunctions";
 
 
-async function impotyWorld(cprovider: CurrentPackProvider, item: PackItem) {
-    const dir = join(item.dir, '..', '..', '..', '..', '..', 'generated', 'minecraft', 'structures');
-    console.log(dir);
-    const items = await window.showQuickPick(readdirSync(dir).map(r => { return { label: r }; }), { canPickMany: true, title: 'Choose structures to import' });
-    if (items?.length === 0) { return notifications.sendErrorMessage('You must choose at least one structure', 'workspace.import.ErrorRequired'); }
+async function impotyWorld(cprovider: CurrentPackProvider, item: PackItem, rootpath: string) {
+    const dir = join(rootpath, '..', '..', 'generated', 'minecraft', 'structures');
+    const items = await window.showQuickPick(UtilFunctions.getFiles(dir, "nbt").map(r => { return { label: r }; }), { canPickMany: true, title: 'Choose structures to import' });
+
+    if (items?.length === 0) { return; }
     assert(items);
-    const errs:NodeJS.ErrnoException[] = [];
+    const errs: NodeJS.ErrnoException[] = [];
     try {
         items.forEach(async ite => {
-            await copyFile(join(dir, ite.label), item.dir, (err) => {
-                if (err) { errs.push(err); };
+            const sdkfng = join(dir, ite.label);
+            const targetDir = join(item.dir,ite.label);
+
+            if (!existsSync(item.dir)) { mkdir(item.dir, { recursive: true }, (e) => { }); };
+
+            
+            const asdfk = targetDir.split("\\");
+            asdfk.pop();
+            if(!existsSync(asdfk.join("\\"))) {mkdirSync(asdfk.join("\\"),{recursive:true});}
+
+            await copyFile(sdkfng, join(item.dir, ite.label), (e) => {
+                if (e) {
+                    notifications.sendErrorMessage("An unknown error occoured:" + e.message, "workspace.import.ErrorUnknown");
+                }
             });
         });
+
+
     } catch (error) {
         notifications.sendErrorMessage(`Couldn't import all structures: ${error}`, 'workspace.import.ErrorUnknown');
     } finally {
@@ -26,19 +41,17 @@ async function impotyWorld(cprovider: CurrentPackProvider, item: PackItem) {
             notifications.sendErrorMessage(`Couldn't import all structures ${errs.join(',')}`, 'workspace.import.ErrorFS');
         } else {
             notifications.sendInformationMessage('Successfully imported all structures!', 'workspace.import.Success');
+            cprovider.refresh();
         }
     }
 
 }
 
 
-export default async (cprovider: CurrentPackProvider, item: PackItem) => {
-    if (item.type === PackItemType.structureRoot) {
-        const from = await window.showQuickPick(["Import from World", "Import from Saved"].map(r => { return { label: r } as QuickPickItem; }), { canPickMany: false, title: 'Choose a directory to import' });
-        if (!from) { return; }
+export default async (cprovider: CurrentPackProvider, item: PackItem, rootpath?: string) => {
+    if (!rootpath) { return; }
 
-        if (from.label === "Import from World") {
-            await impotyWorld(cprovider, item);
-        }
+    if (item.type === PackItemType.structureRoot || item.type === PackItemType.structureFolder) {
+        await impotyWorld(cprovider, item, rootpath);
     }
 };